+++ src/util/random.ts
@@ -139,3 +139,83 @@ export function alphabeticalGenerator(index: number) {
   }
   return name;
 }
+
+export function createZeroWidthGenerator() {
+  var keywords = [
+    "if",
+    "in",
+    "for",
+    "let",
+    "new",
+    "try",
+    "var",
+    "case",
+    "else",
+    "null",
+    "break",
+    "catch",
+    "class",
+    "const",
+    "super",
+    "throw",
+    "while",
+    "yield",
+    "delete",
+    "export",
+    "import",
+    "public",
+    "return",
+    "switch",
+    "default",
+    "finally",
+    "private",
+    "continue",
+    "debugger",
+    "function",
+    "arguments",
+    "protected",
+    "instanceof",
+    "await",
+    "async",
+
+    // new key words and other fun stuff :P
+    "NaN",
+    "undefined",
+    "true",
+    "false",
+    "typeof",
+    "this",
+    "static",
+    "void",
+    "of",
+  ];
+
+  var maxSize = 0;
+  var currentKeyWordsArray: string[] = [];
+
+  function generateArray() {
+    var result = keywords
+      .map(
+        (keyWord) =>
+          keyWord + "\u200C".repeat(Math.max(maxSize - keyWord.length, 1))
+      )
+      .filter((craftedVariableName) => craftedVariableName.length == maxSize);
+
+    if (!result.length) {
+      ++maxSize;
+      return generateArray();
+    }
+
+    return shuffle(result);
+  }
+
+  function getNextVariable(): string {
+    if (!currentKeyWordsArray.length) {
+      ++maxSize;
+      currentKeyWordsArray = generateArray();
+    }
+    return currentKeyWordsArray.pop();
+  }
+
+  return { generate: getNextVariable };
+}